Split EntityAreaTool, EntityNameTool and TextTool
