+++ srcjs/entitystatetool.js
@@ -3,7 +3,6 @@ import { classMap } from 'lit-html/directives/class-map.js';
 import { styleMap } from 'lit-html/directives/style-map.js';
 
 import Merge from './merge';
-import Utils from './utils';
 import BaseTool from './basetool';
 
 /** ****************************************************************************
