+++ srcjs/swiss-army-knife-card.js
@@ -69,8 +69,11 @@ import EllipseTool from './ellipsetool';
 import RectangleTool from './rectangletool';
 import RectangleToolEx from './rectangletoolex';
 import BadgeTool from './badgetool';
-// import EntityIconTool from './entityicontool';
+// import EntityIconTool from './entityicontool'; // Uses SAK as cache. Can't split!
 import EntityStateTool from './entitystatetool';
+import EntityAreaTool from './entityareatool';
+import EntityNameTool from './entitynametool';
+import TextTool from './texttool';
 
 // Original injector is buggy. Use a patched version, and store this local...
 import * as SvgInjector from '../dist/SVGInjector.min.js'; // lgtm[js/unused-local-variable]
@@ -738,285 +741,15 @@ class Toolset {
   ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
  */
 
-/** ****************************************************************************
-  * EntityNameTool class
-  *
-  * Summary.
-  *
-  */
-
-class EntityNameTool extends BaseTool {
-  constructor(argToolset, argConfig, argPos) {
-    const DEFAULT_NAME_CONFIG = {
-      classes: {
-        tool: {
-          'sak-name': true,
-          hover: true,
-        },
-        name: {
-          'sak-name__name': true,
-        },
-      },
-      styles: {
-        tool: {
-        },
-        name: {
-        },
-      },
-    };
-
-    super(argToolset, Merge.mergeDeep(DEFAULT_NAME_CONFIG, argConfig), argPos);
-
-    this._name = {};
-    // Init classes
-    this.classes.tool = {};
-    this.classes.name = {};
-
-    // Init styles
-    this.styles.name = {};
-    if (this.dev.debug) console.log('EntityName constructor coords, dimensions', this.coords, this.dimensions, this.svg, this.config);
-  }
-
-  /** *****************************************************************************
-  * EntityNameTool::_buildName()
-  *
-  * Summary.
-  * Builds the Name string.
-  *
-  */
-
-  _buildName(entityState, entityConfig) {
-    return (
-      this.activeAnimation?.name // Name from animation
-      || entityConfig.name
-      || entityState.attributes.friendly_name
-    );
-  }
-
-  /** *****************************************************************************
-  * EntityNameTool::_renderEntityName()
-  *
-  * Summary.
-  * Renders the entity name using precalculated coordinates and dimensions.
-  * Only the runtime style is calculated before rendering the name
-  *
-  */
-
-  _renderEntityName() {
-    this.MergeAnimationClassIfChanged();
-    this.MergeColorFromState(this.styles.name);
-    this.MergeAnimationStyleIfChanged();
-
-    const name = this.textEllipsis(
-      this._buildName(
-        this._card.entities[this.defaultEntityIndex()],
-        this._card.config.entities[this.defaultEntityIndex()],
-      ),
-      this.config?.show?.ellipsis,
-    );
-
-    return svg`
-        <text>
-          <tspan class="${classMap(this.classes.name)}" x="${this.svg.cx}" y="${this.svg.cy}" style="${styleMap(this.styles.name)}">${name}</tspan>
-        </text>
-      `;
-  }
-
-  /** *****************************************************************************
-  * EntityNameTool::render()
-  *
-  * Summary.
-  * The render() function for this object.
-  *
-  */
-  render() {
-    return svg`
-      <g id="name-${this.toolId}" class="${classMap(this.classes.tool)}"
-        @click=${(e) => this.handleTapEvent(e, this.config)}>
-        ${this._renderEntityName()}
-      </g>
-    `;
-  }
-} // END of class
-
 /**
   ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
   ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
  */
 
-/** ****************************************************************************
-  * EntityAreaTool class
-  *
-  * Summary.
-  *
-  */
-
-class EntityAreaTool extends BaseTool {
-  constructor(argToolset, argConfig, argPos) {
-    const DEFAULT_AREA_CONFIG = {
-      classes: {
-        tool: {
-        },
-        area: {
-          'sak-area__area': true,
-          hover: true,
-        },
-      },
-      styles: {
-        tool: {
-        },
-        area: {
-        },
-      },
-    };
-
-    super(argToolset, Merge.mergeDeep(DEFAULT_AREA_CONFIG, argConfig), argPos);
-
-    // Text is rendered in its own context. No need for SVG coordinates.
-    this.classes.area = {};
-    this.styles.area = {};
-    if (this.dev.debug) console.log('EntityAreaTool constructor coords, dimensions', this.coords, this.dimensions, this.svg, this.config);
-  }
-
-  /** *****************************************************************************
-  * EntityAreaTool::_buildArea()
-  *
-  * Summary.
-  * Builds the Area string.
-  *
-  */
-
-  _buildArea(entityState, entityConfig) {
-    return (
-      entityConfig.area
-      || '?'
-    );
-  }
-
-  /** *****************************************************************************
-  * EntityAreaTool::_renderEntityArea()
-  *
-  * Summary.
-  * Renders the entity area using precalculated coordinates and dimensions.
-  * Only the runtime style is calculated before rendering the area
-  *
-  */
-
-  _renderEntityArea() {
-    this.MergeAnimationClassIfChanged();
-    this.MergeColorFromState(this.styles.area);
-    this.MergeAnimationStyleIfChanged();
-
-    const area = this.textEllipsis(
-      this._buildArea(
-        this._card.entities[this.defaultEntityIndex()],
-        this._card.config.entities[this.defaultEntityIndex()],
-      ),
-      this.config?.show?.ellipsis,
-    );
-
-    return svg`
-        <text>
-          <tspan class="${classMap(this.classes.area)}"
-          x="${this.svg.cx}" y="${this.svg.cy}" style="${styleMap(this.styles.area)}">${area}</tspan>
-        </text>
-      `;
-  }
-
-  /** *****************************************************************************
-  * EntityAreaTool::render()
-  *
-  * Summary.
-  * The render() function for this object.
-  *
-  */
-  render() {
-    return svg`
-      <g id="area-${this.toolId}" class="${classMap(this.classes.tool)}"
-        @click=${(e) => this.handleTapEvent(e, this.config)}>
-        ${this._renderEntityArea()}
-      </g>
-    `;
-  }
-} // END of class
-
 /**
   ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
   ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
  */
-
-/** ****************************************************************************
-  * TextTool class
-  *
-  * Summary.
-  *
-  */
-
-class TextTool extends BaseTool {
-  constructor(argToolset, argConfig, argPos) {
-    const DEFAULT_TEXT_CONFIG = {
-      classes: {
-        tool: {
-          'sak-text': true,
-        },
-        text: {
-          'sak-text__text': true,
-          hover: false,
-        },
-      },
-      styles: {
-        tool: {
-        },
-        text: {
-        },
-      },
-    };
-
-    super(argToolset, Merge.mergeDeep(DEFAULT_TEXT_CONFIG, argConfig), argPos);
-
-    this.EnableHoverForInteraction();
-    this.text = this.config.text;
-    this.styles.text = {};
-    if (this.dev.debug) console.log('TextTool constructor coords, dimensions', this.coords, this.dimensions, this.svg, this.config);
-  }
-
-  /** *****************************************************************************
-  * TextTool::_renderText()
-  *
-  * Summary.
-  * Renders the text using precalculated coordinates and dimensions.
-  * Only the runtime style is calculated before rendering the text
-  *
-  */
-
-  _renderText() {
-    this.MergeAnimationClassIfChanged();
-    this.MergeColorFromState(this.styles.text);
-    this.MergeAnimationStyleIfChanged();
-
-    return svg`
-        <text>
-          <tspan class="${classMap(this.classes.text)}" x="${this.svg.cx}" y="${this.svg.cy}" style="${styleMap(this.styles.text)}">${this.text}</tspan>
-        </text>
-      `;
-  }
-
-  /** *****************************************************************************
-  * TextTool::render()
-  *
-  * Summary.
-  * The render() function for this object.
-  *
-  */
-  render() {
-    return svg`
-        <g id="text-${this.toolId}" class="${classMap(this.classes.tool)}"
-          @click=${(e) => this.handleTapEvent(e, this.config)}>
-          ${this._renderText()}
-        </g>
-      `;
-  }
-} // END of class
 
 /**
   ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
